move relay map and node to base; fix dns stuff
