+++ iroh-base/Cargo.toml
@@ -47,12 +47,13 @@ serde_json = "1.0.107"
 serde_test = "1.0.176"
 
 [features]
-default = ["hash", "base32"]
+default = ["hash", "base32", "relay"]
 hash = ["dep:blake3", "dep:data-encoding", "dep:postcard", "dep:derive_more", "base32"]
 base32 = ["dep:data-encoding", "dep:postcard"]
 redb = ["dep:redb"]
 key = ["dep:ed25519-dalek", "dep:once_cell", "dep:rand", "dep:rand_core", "dep:ssh-key", "dep:ttl_cache", "dep:aead", "dep:crypto_box", "dep:zeroize", "dep:url", "dep:derive_more", "dep:getrandom"]
 wasm = ["getrandom?/js"]
+relay = []
 
 [package.metadata.docs.rs]
 all-features = true
