+++ iroh-base/src/lib.rs
@@ -13,6 +13,10 @@ pub mod key;
 #[cfg(feature = "key")]
 #[cfg_attr(iroh_docsrs, doc(cfg(feature = "key")))]
 pub mod node_addr;
+#[cfg(feature = "relay")]
+#[cfg_attr(iroh_docsrs, doc(cfg(feature = "relay")))]
+pub mod relay_map;
+mod relay_url;
 #[cfg(feature = "base32")]
 #[cfg_attr(iroh_docsrs, doc(cfg(feature = "base32")))]
 pub mod ticket;
