+++ iroh-netcheck/src/dns.rs
@@ -201,12 +201,68 @@ async fn stagger_call<T, F: Fn() -> Fut, Fut: Future<Output = Result<T>>>(
 
 #[cfg(test)]
 pub(crate) mod tests {
+    use std::net::Ipv6Addr;
     use std::sync::atomic::AtomicUsize;
     use std::time::Duration;
+
+    use once_cell::sync::Lazy;
 
     use super::*;
     const TIMEOUT: Duration = Duration::from_secs(5);
     const STAGGERING_DELAYS: &[u64] = &[200, 300];
+
+    static DNS_RESOLVER: Lazy<TokioAsyncResolver> =
+        Lazy::new(|| create_default_resolver().expect("unable to create DNS resolver"));
+
+    /// Get the DNS resolver used within iroh-net.
+    pub fn resolver() -> &'static TokioAsyncResolver {
+        Lazy::force(&DNS_RESOLVER)
+    }
+
+    /// Deprecated IPv6 site-local anycast addresses still configured by windows.
+    ///
+    /// Windows still configures these site-local addresses as soon even as an IPv6 loopback
+    /// interface is configured.  We do not want to use these DNS servers, the chances of them
+    /// being usable are almost always close to zero, while the chance of DNS configuration
+    /// **only** relying on these servers and not also being configured normally are also almost
+    /// zero.  The chance of the DNS resolver accidentally trying one of these and taking a
+    /// bunch of timeouts to figure out they're no good are on the other hand very high.
+    const WINDOWS_BAD_SITE_LOCAL_DNS_SERVERS: [IpAddr; 3] = [
+        IpAddr::V6(Ipv6Addr::new(0xfec0, 0, 0, 0xffff, 0, 0, 0, 1)),
+        IpAddr::V6(Ipv6Addr::new(0xfec0, 0, 0, 0xffff, 0, 0, 0, 2)),
+        IpAddr::V6(Ipv6Addr::new(0xfec0, 0, 0, 0xffff, 0, 0, 0, 3)),
+    ];
+
+    /// Get resolver to query MX records.
+    ///
+    /// We first try to read the system's resolver from `/etc/resolv.conf`.
+    /// This does not work at least on some Androids, therefore we fallback
+    /// to the default `ResolverConfig` which uses eg. to google's `8.8.8.8` or `8.8.4.4`.
+    fn create_default_resolver() -> Result<TokioAsyncResolver> {
+        let (system_config, mut options) =
+            hickory_resolver::system_conf::read_system_conf().unwrap_or_default();
+
+        // Copy all of the system config, but strip the bad windows nameservers.  Unfortunately
+        // there is no easy way to do this.
+        let mut config = hickory_resolver::config::ResolverConfig::new();
+        if let Some(name) = system_config.domain() {
+            config.set_domain(name.clone());
+        }
+        for name in system_config.search() {
+            config.add_search(name.clone());
+        }
+        for nameserver_cfg in system_config.name_servers() {
+            if !WINDOWS_BAD_SITE_LOCAL_DNS_SERVERS.contains(&nameserver_cfg.socket_addr.ip()) {
+                config.add_name_server(nameserver_cfg.clone());
+            }
+        }
+
+        // see [`ResolverExt::lookup_ipv4_ipv6`] for info on why we avoid `LookupIpStrategy::Ipv4AndIpv6`
+        options.ip_strategy = hickory_resolver::config::LookupIpStrategy::Ipv4thenIpv6;
+
+        let resolver = hickory_resolver::AsyncResolver::tokio(config, options);
+        Ok(resolver)
+    }
 
     #[tokio::test]
     async fn stagger_basic() {
